sampels

+++ src/main/java/com/riskified/models/RefundDetails.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public class RefundDetails {
   public String refundId;
   public Date refundedAt;
-  public Float amount;
+  public double amount;
   public String currency;
   public String reason;
 }
